+++ src/controls/dynamicForm/IDynamicFormProps.ts
@@ -1,5 +1,6 @@
 import { ExtensionContext } from '@microsoft/sp-extension-base';
 import { WebPartContext } from '@microsoft/sp-webpart-base';
+import { IItem } from '@pnp/sp/items';
 
 export interface IDynamicFormProps {
   /**
@@ -22,7 +23,7 @@ export interface IDynamicFormProps {
   /**
    * Handler for form submitted event
    */
-  onSubmitted?: (listItemData: any) => void;
+  onSubmitted?: (listItemData: any, listItem?: IItem) => void;
   /**
    * Handler of submission error
    */
@@ -39,6 +40,12 @@ export interface IDynamicFormProps {
    * Content type id of the item
    */
   contentTypeId?: string;
+
+  /**
+   * Specifies if onSubmitted event should pass PnPJS list item (IItem) as a second parameter. Default - true
+   */
+  returnListItemInstanceOnSubmit?: boolean;
+
   /**
  * Used to execute WebSearch. If not provided SearchTab will not be available.
  */