+++ src/controls/dynamicForm/dynamicField/DynamicField.tsx
@@ -74,7 +74,9 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
       isRichText,
       //bingAPIKey,
       dateFormat,
-      columnInternalName
+      columnInternalName,
+      principalType,
+      description
     } = this.props;
 
     const {
@@ -94,6 +96,7 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
     const labelEl = <label className={(required) ? styles.fieldRequired + ' ' + styles.fieldLabel : styles.fieldLabel}>{labelText}</label>;
     const errorText = this.getRequiredErrorText();
     const errorTextEl = <text className={styles.errormessage}>{errorText}</text>;
+    const descriptionEl = <text className={styles.fieldDescription}>{description}</text>;
 
     switch (fieldType) {
       case 'loading':
@@ -118,6 +121,7 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
             onBlur={this.onBlur}
             errorMessage={errorText}
           />
+          {descriptionEl}
         </div>;
 
       case 'Note':
@@ -134,6 +138,7 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
               onChange={(newText) => { this.onChange(newText); return newText; }}
               isEditMode={disabled}>
             </RichText>
+            {descriptionEl}
             {errorTextEl}
           </div>;
         }
@@ -153,6 +158,7 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
               onBlur={this.onBlur}
               errorMessage={errorText}
             />
+            {descriptionEl}
           </div>;
         }
 
@@ -168,6 +174,7 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
             onChange={(e, option) => { this.onChange(option); }}
             onBlur={this.onBlur}
             errorMessage={errorText} />
+          {descriptionEl}
         </div>;
 
       case 'MultiChoice':
@@ -183,6 +190,7 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
             multiSelect
             onBlur={this.onBlur}
             errorMessage={errorText} />
+          {descriptionEl}
         </div>;
 
       case 'Location':
@@ -199,6 +207,7 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
             defaultValue={defaultValue}
             errorMessage={errorText}
           />
+          {descriptionEl}
         </div>;
 
       case 'Lookup':
@@ -218,6 +227,7 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
             onSelectedItem={(newValue) => { this.onChange(newValue); }}
             context={context}
           />
+          {descriptionEl}
           {errorTextEl}
         </div>;
 
@@ -238,6 +248,7 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
             onSelectedItem={(newValue) => { this.onChange(newValue); }}
             context={context}
           />
+          {descriptionEl}
           {errorTextEl}
         </div>;
 
@@ -256,6 +267,7 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
             disabled={disabled}
             onBlur={this.onBlur}
             errorMessage={errorText} />
+          {descriptionEl}
         </div>;
 
       case 'Currency':
@@ -273,6 +285,7 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
             disabled={disabled}
             onBlur={this.onBlur}
             errorMessage={errorText} />
+          {descriptionEl}
         </div>;
 
       case 'DateTime':
@@ -301,6 +314,7 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
               onChange={(newDate) => { this.onChange(newDate); }}
               disabled={disabled} />
           }
+          {descriptionEl}
           {errorTextEl}
         </div>;
 
@@ -318,6 +332,7 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
             onChange={(e, checkedvalue) => { this.onChange(checkedvalue); }}
             disabled={disabled}
           />
+          {descriptionEl}
           {errorTextEl}
         </div>;
 
@@ -335,11 +350,12 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
             personSelectionLimit={1}
             showtooltip={false}
             showHiddenInUI={false}
-            principalTypes={[PrincipalType.User]} // TODO: principal types should be read from the column settings
+            principalTypes={principalType === 'PeopleOnly' ? [PrincipalType.User] : [PrincipalType.User, PrincipalType.SharePointGroup, PrincipalType.DistributionList, PrincipalType.SecurityGroup]}
             resolveDelay={1000}
             onChange={(items) => { this.onChange(items); }}
             disabled={disabled}
           />
+          {descriptionEl}
           {errorTextEl}
         </div>;
 
@@ -357,11 +373,12 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
             personSelectionLimit={30}
             showtooltip={false}
             showHiddenInUI={false}
-            principalTypes={[PrincipalType.User]} // TODO: principal types should be read from the column settings
+            principalTypes={principalType === 'PeopleOnly' ? [PrincipalType.User] : [PrincipalType.User, PrincipalType.SharePointGroup, PrincipalType.DistributionList, PrincipalType.SecurityGroup]}
             resolveDelay={1000}
             onChange={(items) => { this.onChange(items); }}
             disabled={disabled}
           />
+          {descriptionEl}
           {errorTextEl}
         </div>;
 
@@ -387,6 +404,7 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
               onChange={(e, newText) => { this.onURLChange(newText, false); }}
               disabled={disabled} />
           </Stack>
+          {descriptionEl}
           {errorTextEl}
         </div>;
 
@@ -434,6 +452,7 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
                 />}
             </div>
           </Stack>
+          {descriptionEl}
           {errorTextEl}
         </div>;
 
@@ -457,6 +476,7 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
               isTermSetSelectable={false}
             />
           </div>
+          {descriptionEl}
           {errorTextEl}
         </div>;
 
@@ -479,6 +499,7 @@ export class DynamicField extends React.Component<IDynamicFieldProps, IDynamicFi
               onChange={(newValue?: IPickerTerms) => { this.onChange(newValue); }}
               isTermSetSelectable={false} />
           </div>
+          {descriptionEl}
           {errorTextEl}
         </div>;
     }