+++ src/webparts/controlsTest/components/IControlsTestProps.ts
@@ -29,10 +29,12 @@ export interface IControlsTestState {
   canMovePrev: boolean;
   canMoveNext: boolean;
   comboBoxListItemPickerListId: string;
+  comboBoxListItemPickerIds: any[];
   filePickerResult?: IFilePickerResult[];
   treeViewSelectedKeys?: string[];
   showAnimatedDialog?: boolean;
   showCustomisedAnimatedDialog?: boolean;
   showSuccessDialog?: boolean;
   showErrorDialog?: boolean;
+  filePickerDefaultFolderAbsolutePath?: string;
 }